modify_hampelfilter_location & code
